+++ core/CHANGELOG.md
@@ -26,6 +26,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fix red screen on shutdown.  [#1658]
 - Empty passphrase is properly cached in Cardano functions  [#1659]
 
+### Security
+- Ensure that all testnet coins use SLIP-44 coin type 1.
+- Disable all testnet coins from accessing Bitcoin paths.
+- Restrict BIP-45 paths to Bitcoin and coins with strong replay protection.
+- Fix operation source account encoding in Stellar.
+
 
 ## 2.4.0 [9th June 2021]