+++ legacy/firmware/CHANGELOG.md
@@ -10,6 +10,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Removed support for Firo  [#1647]
 - Removed support for Hatch  [#1650]
 
+### Fixed
+- Allow non-standard paths used by Unchained Capital, Green Address and Casa.  [#1660]
+
+### Security
+- Ensure that all testnet coins use SLIP-44 coin type 1.
+- Restrict BIP-45 paths to Bitcoin and coins with strong replay protection.
+- Don't show addresses that have an unrecognized path.
+- Disable all testnet coins from accessing Bitcoin paths.
+- Restrict the BIP-32 path ranges of `account`, `change` and `address_index` fields.
+- Fix operation source account encoding in Stellar.
+
 
 ## 1.10.1 [9th June 2021]
 
@@ -400,3 +411,4 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 [#1627]: https://github.com/trezor/trezor-firmware/issues/1627
 [#1647]: https://github.com/trezor/trezor-firmware/issues/1647
 [#1650]: https://github.com/trezor/trezor-firmware/issues/1650
+[#1660]: https://github.com/trezor/trezor-firmware/issues/1660